fix(academic): 修复教务只读回归并解耦实现
